+++ src/MathParser.h
@@ -102,7 +102,7 @@ private:
   int m_ParserStyle;
   int m_FracStyle;
   //! The maximum number of digits of a number that is to be displayed
-  int m_displayedDigits;
+  long m_displayedDigits;
   bool m_highlight;
   wxFileSystem *m_fileSystem; // used for loading pictures in <img> and <slide>
 };
